blockchain, sizehelper: move sizehelper to separate package
This is done so that some of the code in package blockchain can be
separated out into other packages to better guarntees that there isn't
unexported code that is being accessed by a function that shouldn't be
accessing it.
